函数参数修订，编译命令：nvcc -rdc=true calculate_gpu.cu -o calculate_gpu

+++ calculate.cuh
@@ -4,7 +4,7 @@ __device__ __forceinline__ void calculate_HUV
 const int jt, const int NHQ, const double C0, const double C1, const double HM1, const double HM2, const int DT, const double QLUA, const double VMIN,
 int MBQ_LEN, int MBZQ_LEN, int MBZ_LEN, int MBW_LEN, int MDI_LEN,//数组长度
 int* NV, double* H_pre, double* U_pre, double* V_pre, double* Z_pre, double* MBQ, double* DQT, double* MBZQ, double* ZB1, double* ZBC, double* MBZ, double* DZT, double* TOPW, double* MBW, double* TOPD, double* MDI, double* AREA, double* FNC,
-double** KLAS, double** QT, double** SIDE, double** ZW, double** QW, double** ZT, double** NAC, double** COSF, double** SINF, double** SLCOS, double** SLSIN,
+double** KLAS, double** QT, double** SIDE, double** ZW, double** QW, double** ZT, int** NAC, double** COSF, double** SINF, double** SLCOS, double** SLSIN,
 double* H_res, double* U_res, double* V_res, double* Z_res, double* W_res);
 
 __device__ __forceinline__ void calculate_WHUV
@@ -12,21 +12,21 @@ __device__ __forceinline__ void calculate_WHUV
 const int jt, const int NHQ, const double C0, const double C1, const double HM1, const double HM2, const double VMIN,
 int MBQ_LEN, int MBZQ_LEN, int MBZ_LEN, int MBW_LEN, int MDI_LEN,//数组长度
 double* H_pre, double* U_pre, double* V_pre, double* Z_pre, double* MBQ, double* DQT, double* MBZQ, double* ZB1, double* ZBC, double* MBZ, double* DZT, double* TOPW, double* MBW, double* TOPD, double* MDI,
-double** KLAS, double** QT, double** SIDE, double** ZW, double** QW, double** ZT, double** NAC, double** COSF, double** SINF, double** SLCOS, double** SLSIN);
+double** KLAS, double** QT, double** SIDE, double** ZW, double** QW, double** ZT, int** NAC, double** COSF, double** SINF, double** SLCOS, double** SLSIN);
 
 __device__ __forceinline__ void calculate_FLUX
 (int t, int pos, double (&FLUX)[4][4],
 const int jt, const int NHQ, const double C0, const double C1, const double HM1, const double HM2, const double VMIN,
 int MBQ_LEN, int MBZQ_LEN, int MBZ_LEN, int MBW_LEN, int MDI_LEN,//数组长度
 double* H_pre, double* U_pre, double* V_pre, double* Z_pre, double* MBQ, double* DQT, double* MBZQ, double* ZB1, double* ZBC, double* MBZ, double* DZT, double* TOPW, double* MBW, double* TOPD, double* MDI,
-double** KLAS, double** QT, double** SIDE, double** ZW, double** QW, double** ZT, double** NAC, double** COSF, double** SINF);
+double** KLAS, double** QT, double** SIDE, double** ZW, double** QW, double** ZT, int** NAC, double** COSF, double** SINF);
 
 __device__ __forceinline__ void BOUNDA
 (int t, int j, int pos, double (&FLUX)[4][4],double (&QL)[3],double (&QR)[3], double &FIL, double HC, double UC, double VC, double ZC,
 const int jt, const int NHQ, const double C0, const double C1,
 int MBQ_LEN, int MBZQ_LEN, int MBZ_LEN, int MBW_LEN, int MDI_LEN,//数组长度
 double* H_pre, double* Z_pre, double* MBQ, double* DQT, double* MBZQ, double* ZBC, double* MBZ, double* DZT, double* TOPW, double* MBW, double* TOPD, double* MDI,
-double** KLAS, double** QT, double** SIDE, double** ZW, double** QW, double** ZT, double** NAC, double** COSF, double** SINF);
+double** KLAS, double** QT, double** SIDE, double** ZW, double** QW, double** ZT, int** NAC, double** COSF, double** SINF);
 
 __device__ __forceinline__ void OSHER
 (int t, int pos, double (&QL)[3], double (&QR)[3], double &FIL, double (&FLR_OSHER)[4],
